Employee -UserId Match
This program extracts data from the db in order to match employees with
their userid , in order to determine their employment status with the
objective of knowing who will need a new update for an inhouse
application. The program extracts data from the db , matches the userid
with employee record on table taking into account a username pattern .
Once it's matched it create an excel sheet with the necessary
information.